black: enable python target-version autodetection
Since 23.1.0, black supports autodetecting the target version of python
for which it is supposed to format. This is detected either by
inspecting pyproject.toml files of the project, orfalling back onto a
file-by-file decision.
Removing fixed target versions from our pre-commit template avoids us
having to constantly adjust the versions. One case where it used to fail
was the usage of pattern matching in python-3.10 code, but an older
target-version.

If you need to rely on black implicitly checking for syntax
compatibility with a certain python version, re-add a
[tool.black].target-version config to your project's pyproject.toml.